requirements.txt
aiogram==3.22.0
aiosqlite==0.22.1
apscheduler==3.11.1
Pillow==12.0.0
psutil==7.0.0
pydantic==2.11.5
Pyrogram==2.0.106
